+++ crates/version-spec/tests/resolved_spec_test.rs
@@ -1,5 +1,5 @@
 use semver::Version;
-use version_spec::VersionSpec;
+use version_spec::{CalVer, SemVer, VersionSpec};
 
 mod resolved_spec {
     use super::*;
@@ -33,28 +33,38 @@ mod resolved_spec {
     fn versions() {
         assert_eq!(
             VersionSpec::parse("v1.2.3").unwrap(),
-            VersionSpec::Version(Version::new(1, 2, 3))
+            VersionSpec::Semantic(SemVer(Version::new(1, 2, 3)))
         );
         assert_eq!(
             VersionSpec::parse("1.2.3").unwrap(),
-            VersionSpec::Version(Version::new(1, 2, 3))
+            VersionSpec::Semantic(SemVer(Version::new(1, 2, 3)))
+        );
+        assert_eq!(
+            VersionSpec::parse("1.2.3-0").unwrap(),
+            VersionSpec::Semantic(SemVer(Version::parse("1.2.3-0").unwrap()))
+        );
+        assert_eq!(
+            VersionSpec::parse("1.2.3-alpha").unwrap(),
+            VersionSpec::Semantic(SemVer(Version::parse("1.2.3-alpha").unwrap()))
+        );
+        assert_eq!(
+            VersionSpec::parse("1.2.3-alpha.1").unwrap(),
+            VersionSpec::Semantic(SemVer(Version::parse("1.2.3-alpha.1").unwrap()))
+        );
+
+        // calver
+        assert_eq!(
+            VersionSpec::parse("2024-02").unwrap(),
+            VersionSpec::Calendar(CalVer(Version::new(2024, 2, 0)))
+        );
+        assert_eq!(
+            VersionSpec::parse("2024-2-26").unwrap(),
+            VersionSpec::Calendar(CalVer(Version::new(2024, 2, 26)))
         );
     }
 
     #[test]
-    #[should_panic(expected = "unexpected end of input while parsing minor version number")]
-    fn error_missing_patch() {
-        VersionSpec::parse("1.2").unwrap();
-    }
-
-    #[test]
-    #[should_panic(expected = "unexpected end of input while parsing major version number")]
-    fn error_missing_minor() {
-        VersionSpec::parse("1").unwrap();
-    }
-
-    #[test]
-    #[should_panic(expected = "unexpected character '%' while parsing major version number")]
+    #[should_panic(expected = "ResolvedUnknownFormat")]
     fn error_invalid_char() {
         VersionSpec::parse("%").unwrap();
     }